Implements code that runs regression models
The commit primarily contains eda of the dataframes built with the code in the second commit (02_).  The eda in script 03_ contains a series of multiple variable regressions.  The regression progress using a backward variable identification method to identify the primary variables influence profit.  The first regression contains all variables, then each subsequent regression removes the variable with the highest p value. Each regression has a “fit” and an error normalcy test.  The normally test, test to see if the errors are normally distrusted.   The fit v. residuals test for constant variation in the errors.

Code in 02_ contains some adjustment to the observation in the dataframes such as converting NAs to 0 to count family sizes.

gitignore was updated with .pdf