Implements code that builds and evaluates household and land influences
The major change is script 7.  Script 7 merges the household and land dataframes into a capstone dataframe that is then run through various models with profit as the dependent variables.

The models’-built start with the top variables drawn from the household and land analysis.  Following the initial descriptive statistics and correlation review, the models begin by alternating the removal of variables or the inclusion of quadratics.  The models are also followed by ggplot calls to assess error distributions and residual v. fitted assessments.

Following the last model, there are a series of code that looks specifically at the model’s quadratics, visualizing how the coefficients change per changes in the x values.

A minor change adds the library calls to each script.

The final changes are mainly to cleanup in the visualization folder.